Add support for array_map smartmodule with test suite
The test creates a topic, and sends a series of records containing
JSON which looks like:

```json
{ "importantNumbers": [1, 2, 3, ...] }
```

The hardcoded WASM test module in array_map_code.js uses serde to
extract that field, and emits each of the members as a new Record.

The test creates a consumer with that module and verifies the output.